refactor model and checkpointing
